Refactor separating classes
